US106353 - move sorting to controller
