+++ test/d2l-quick-eval/d2l-quick-eval-activities-list-sorting.html
@@ -12,12 +12,12 @@
 		<script src="../../../lie/dist/lie.polyfill.min.js"></script>
 		<script type="module" src="../../../whatwg-fetch/fetch.js"></script>
 		<script type="module" src="../../../url-polyfill/url-polyfill.js"></script>
-		<script type="module" src="../../components/d2l-quick-eval/d2l-quick-eval-activities-list.js"></script>
+		<script type="module" src="../../components/d2l-quick-eval/d2l-quick-eval-submissions.js"></script>
 	</head>
 	<body>
 		<test-fixture id="basic">
 			<template strip-whitespace>
-				<d2l-quick-eval-activities-list></d2l-quick-eval-activities-list>
+				<d2l-quick-eval-submissions></d2l-quick-eval-submissions>
 			</template>
 		</test-fixture>
